fixed the issue with reboot ovn compute node
ovn compute node needs public nic to be added to external bridge
manually. If the changes were not made into the interfaces, then
the compute node can not survice the machine reboot. With this
fix, the ovn compute node can be restarted/rebooted and the
external bridge will be automatically configured.